backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymilvus==2.4.4
python-dotenv==1.0.1
google-api-python-client==2.126.0
stockfish
langgraph
python-chess